fix setup script

+++ setup.py
@@ -1,11 +1,13 @@
 #!/usr/bin/env python3
 
+import os
 import re
 import shutil
 
 from setuptools import setup, find_packages
 
-shutil.copyfile('buku', 'buku.py')
+if os.path.isfile('buku'):
+    shutil.copyfile('buku', 'buku.py')
 
 with open('buku.py', encoding='utf-8') as f:
     version = re.search('__version__ = \'([^\']+)\'', f.read()).group(1)
